requirements.txt
pyserial
pyyaml
paho-mqtt